requirements.txt
tqdm
pyyaml
numpy<2
resampy
scipy
pandas
matplotlib
torch
pytorch-lightning
tensorboard
torchinfo
torchaudio
torchmetrics
scikit_learn
#git+https://github.com/Blealtan/efficient-kan.git
essentia
librosa
soundfile
yt-dlp
certifi
brotli
websockets
requests
curl_cffi
mutagen
xattr
pycryptodome
pycryptodomex
SecretStorage
